Skip check for cached position in ChunkRetriever when not passed as arg
The only way ChunkRetriever might have cached a position that has been
undefined in the redux store (from where it is passed to getNextChunk by
App) is for the fetcher to have returned undefined for the new chunk id.

Then, the cached position in the store will be updated with an
'undefined', but the ChunkRetriever will store the "repaired" value of
one, because it will increment the cached position (undefined so far)
with 1.