Start shifting to a unified way of storing text sources
Created a PersistentTextSource in store.ts, which will eventually
replace LocalTextSource. Use PTS in the ChunkRetriever to unify the way
local and remote text sources are handled.

Source fetcher creators now accept only a single string argument from
which all source information needs to be infered. Memoizing the
functions in the sourceFetchers map helps with not splitting a local
source text to an array of its lines, for example.